Consolidate Vega actions.
